requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
python-multipart==0.0.6
streamlit==1.28.2